getter 반환 타입 수정

+++ src/io/npee/designpatterns/_02_observer/_01_base/WeatherData.java
@@ -13,15 +13,15 @@ public class WeatherData {
 		// to-do
 	}
 
-	public double getTemperature() {
+	public float getTemperature() {
 		return temperature;
 	}
 
-	public double getHumidity() {
+	public float getHumidity() {
 		return humidity;
 	}
 
-	public double getPressure() {
+	public float getPressure() {
 		return pressure;
 	}
 }
