+++ src/io/npee/designpatterns/_02_observer/_03_using_java_utils/WeatherData.java
@@ -28,15 +28,15 @@ public class WeatherData extends Observable {
 		measurementsChanged();
 	}
 
-	public double getTemperature() {
+	public float getTemperature() {
 		return temperature;
 	}
 
-	public double getHumidity() {
+	public float getHumidity() {
 		return humidity;
 	}
 
-	public double getPressure() {
+	public float getPressure() {
 		return pressure;
 	}
 }
